feat: renomear exemplos do módulo dois para padronizar nomenclatura
