+++ src/pages/ModuleTwo/Exemplos/ExemploTres.tsx
@@ -10,7 +10,7 @@ const buscaItens = (busca: string, itens: Item[]) => {
   return itens.filter(item => item.nome.toLowerCase().includes(busca.toLowerCase()));
 };
 
-function BuscaDeItens() {
+function ExemploTres() {
   // Lista de itens (simulando uma lista grande de dados)
   const itens: Item[] = useMemo(() => {
     return Array.from({ length: 10000 }, (_, indice) => ({ nome: `Item ${indice}` }));
@@ -52,4 +52,4 @@ function BuscaDeItens() {
   );
 }
 
-export default BuscaDeItens;
+export default ExemploTres;
